Switch build from gcc to clang-20 and add lint rules for all C code
Use clang-20 instead of gcc for compiling both CPU and GPU backends.
Add dune lint aliases that run clang-20 -fsyntax-only with -Weverything
(matching .clangd flags) for cpu/, gpu/, shared/, and utilities/.
Fix warnings: strict prototype in lux_cpu.h, sign conversions in
cuda_helpers.c, and struct padding in cuda_helpers.h.

+++ utilities/cuda_helpers.c
@@ -8,8 +8,8 @@ char* read_file(const char* path) {
   fseek(f, 0, SEEK_END);
   long len = ftell(f);
   fseek(f, 0, SEEK_SET);
-  char* buf = malloc(len + 1);
-  size_t n = fread(buf, 1, len, f);
+  char* buf = malloc((size_t)len + 1);
+  size_t n = fread(buf, 1, (size_t)len, f);
   buf[n] = '\0';
   fclose(f);
   return buf;
